hunting the impossible bug

+++ crates/golgge/src/simplify_and.rs
@@ -45,8 +45,8 @@ pub trait WithAnd: WithTrue {
         for i in 1..(n as egg::uvar) {
             ret.push(ENodeOrVar::Var(Var::from_usize(from + i)));
             ret.push(ENodeOrVar::ENode(Self::mk_and(
-                (2 * (i - 1) as usize).into(),
-                ((2 * (i - 1)) + 1 as usize).into(),
+                (2 * (i - 1)).into(),
+                ((2 * (i - 1)) + 1_usize).into(),
             )));
         }
         ret.into()
